feat(app): updated file paths, moved unused mockups into separate folder
