+++ src/components/mockups/game-queue-player/index.jsx
