+++ src/components/mockups/main-screen/main-screen.css
